Move libcurl stuff into networking layer
