+++ GPTifier/include/networking/curl_base.hpp
