+++ GPTifier/src/command_fine_tune.cpp
@@ -1,9 +1,9 @@
 #include "command_fine_tune.hpp"
 
-#include "api_openai_user.hpp"
 #include "cli.hpp"
 #include "help_messages.hpp"
 #include "models.hpp"
+#include "networking/api_openai_user.hpp"
 #include "params.hpp"
 #include "parsers.hpp"
 #include "utils.hpp"
