+++ GPTifier/src/command_run.cpp
@@ -1,9 +1,9 @@
 #include "command_run.hpp"
 
-#include "api_openai_user.hpp"
 #include "cli.hpp"
 #include "datadir.hpp"
 #include "models.hpp"
+#include "networking/api_openai_user.hpp"
 #include "params.hpp"
 #include "parsers.hpp"
 #include "selectors.hpp"
