+++ GPTifier/src/networking/api_openai_user.cpp
@@ -1,4 +1,4 @@
-#include "api_openai_user.hpp"
+#include "networking/api_openai_user.hpp"
 
 #include "configs.hpp"
