+++ GPTifier/src/networking/curl_base.cpp
@@ -1,4 +1,4 @@
-#include "curl_base.hpp"
+#include "networking/curl_base.hpp"
 
 #include <stdexcept>
